feat: add Vitest configuration and comprehensive test suite for TypeScript utilities
- Introduced a new Vitest configuration file to manage test coverage and settings.
- Added extensive test cases for various TypeScript utility functions, including `genClass`, `genConstructor`, `genProperty`, `genSetter`, and control flow utilities like `genIf`, `genSwitch`, and `genWhile`.
- Implemented tests for decorators, enums, and conditional types, enhancing overall test coverage and ensuring accurate output for generated code.
- Organized tests into dedicated files for better maintainability and clarity.
