+++ test/interface.test.ts
@@ -1,0 +1,195 @@
+import { expect, describe, it } from "vitest";
+import { genInterface } from "../src";
+import { genTestTitle } from "./_utils";
+
+const genInterfaceTests: Array<{
+  input: Parameters<typeof genInterface>;
+  code: string;
+}> = [
+  { input: ["FooInterface"], code: "interface FooInterface {}" },
+  {
+    input: ["FooInterface", undefined, { extends: ["Other"] }],
+    code: "interface FooInterface extends Other {}",
+  },
+  {
+    input: ["FooInterface", undefined, { extends: "Other" }],
+    code: "interface FooInterface extends Other {}",
+  },
+  {
+    input: [
+      "FooInterface",
+      { name: "boolean", 'other name"': { value: "() => {}" } },
+    ],
+    code: `interface FooInterface {
+  name: boolean
+  "other name\\"": {
+    value: () => {}
+  }
+}`,
+  },
+  {
+    input: ["FooInterface", { "na'me?": "boolean" }],
+    code: `interface FooInterface {
+  "na'me"?: boolean
+}`,
+  },
+  {
+    input: ["FooInterface", {}, { jsdoc: "Simple description" }],
+    code: `/** Simple description */
+interface FooInterface {}`,
+  },
+  {
+    input: [
+      "FooInterface",
+      {},
+      {
+        jsdoc: ["Complex description", "@param someParam", "@returns void"],
+      },
+    ],
+    code: `/**
+ * Complex description
+ * @param someParam
+ * @returns void
+ */
+interface FooInterface {}`,
+  },
+  {
+    input: [
+      "FooInterface",
+      {
+        prop: {
+          type: "string",
+          jsdoc: "Property description",
+        },
+      },
+    ],
+    code: `interface FooInterface {
+  /** Property description */
+  prop: string
+}`,
+  },
+  {
+    input: [
+      "FooInterface",
+      {
+        prop: {
+          type: "string",
+          jsdoc: {
+            description: "Complex prop",
+            default: "''",
+            deprecated: "use newProp instead",
+          },
+        },
+      },
+    ],
+    code: `interface FooInterface {
+  /**
+   * Complex prop
+   * @default ''
+   * @deprecated use newProp instead
+   */
+  prop: string
+}`,
+  },
+  {
+    input: [
+      "FooInterface",
+      {
+        nested: {
+          subProp: {
+            type: "boolean",
+            jsdoc: "Nested property",
+          },
+        },
+      },
+    ],
+    code: `interface FooInterface {
+  nested: {
+    /** Nested property */
+    subProp: boolean
+  }
+}`,
+  },
+  {
+    input: [
+      "FooInterface",
+      {
+        nested: {
+          subProp: {
+            type: "boolean",
+            jsdoc: "Nested property",
+          },
+        },
+      },
+      {
+        export: true,
+      },
+    ],
+    code: `export interface FooInterface {
+  nested: {
+    /** Nested property */
+    subProp: boolean
+  }
+}`,
+  },
+  {
+    input: [
+      "FooInterface",
+      {
+        nested: {
+          subProp: {
+            type: "boolean",
+            jsdoc: "Nested property",
+          },
+        },
+      },
+      {
+        export: true,
+        jsdoc: ["Complex description", "@param someParam", "@returns void"],
+      },
+    ],
+    code: `/**
+ * Complex description
+ * @param someParam
+ * @returns void
+ */
+export interface FooInterface {
+  nested: {
+    /** Nested property */
+    subProp: boolean
+  }
+}`,
+  },
+  {
+    input: [
+      "FooInterface",
+      [
+        { name: "foo", type: "string" },
+        { name: "bar", type: "number", optional: true },
+      ],
+    ],
+    code: `interface FooInterface {
+  foo: string
+  bar?: number
+}`,
+  },
+  {
+    input: [
+      "FooInterface",
+      [{ name: "id", type: "string", jsdoc: "Unique id" }],
+    ],
+    code: `interface FooInterface {
+  /** Unique id */
+  id: string
+}`,
+  },
+];
+
+describe("genInterface", () => {
+  for (const t of genInterfaceTests) {
+    it(genTestTitle(t.code), () => {
+      const code = genInterface(...t.input);
+      expect(code).to.equal(t.code);
+    });
+  }
+});
