+++ test/object.test.ts
@@ -1,5 +1,12 @@
 import { expect, describe, it } from "vitest";
-import { genObject, genLiteral, genMap, genSet } from "../src";
+import {
+  genObject,
+  genLiteral,
+  genMap,
+  genSet,
+  genGetter,
+  genMethod,
+} from "../src";
 import { genTestTitle } from "./_utils";
 
 const genObjectTests = [
@@ -184,5 +191,71 @@ describe("genSet", () => {
         [3, 4],
       ]),
     ).to.equal("new Set([\n[\n    1,\n    2\n  ],\n[\n    3,\n    4\n  ]\n])");
+  });
+});
+
+describe("genGetter", () => {
+  it("get value() { return this._v; }", () => {
+    expect(genGetter("value", ["return this._v;"])).to.equal(
+      "get value() {\n  return this._v;\n}",
+    );
+  });
+  it("get id(): string { return this._id; }", () => {
+    expect(
+      genGetter("id", ["return this._id;"], { returnType: "string" }),
+    ).to.equal("get id(): string {\n  return this._id;\n}");
+  });
+  it("getter with JSDoc", () => {
+    expect(
+      genGetter("x", ["return this._x;"], { jsdoc: "Getter for x" }),
+    ).to.equal("/** Getter for x */\nget x() {\n  return this._x;\n}");
+  });
+  it("getter with empty body", () => {
+    expect(genGetter("empty", [])).to.equal("get empty() {}");
+  });
+});
+
+describe("genMethod (kind: get/set)", () => {
+  it("get value(): number { return this._v; }", () => {
+    expect(
+      genMethod({
+        name: "value",
+        kind: "get",
+        body: ["return this._v;"],
+        returnType: "number",
+      }),
+    ).to.equal("get value(): number {\n  return this._v;\n}");
+  });
+  it("set value(v: number) { this._v = v; }", () => {
+    expect(
+      genMethod({
+        name: "value",
+        kind: "set",
+        parameters: [{ name: "v", type: "number" }],
+        body: ["this._v = v;"],
+      }),
+    ).to.equal("set value(v: number) {\n  this._v = v;\n}");
+  });
+  it("static get id() with static getter", () => {
+    expect(
+      genMethod({
+        name: "id",
+        kind: "get",
+        static: true,
+        body: ["return this._id;"],
+        returnType: "string",
+      }),
+    ).to.equal("static get id(): string {\n  return this._id;\n}");
+  });
+  it("static set value(v) with static setter", () => {
+    expect(
+      genMethod({
+        name: "value",
+        kind: "set",
+        static: true,
+        parameters: [{ name: "v", type: "number" }],
+        body: ["this._v = v;"],
+      }),
+    ).to.equal("static set value(v: number) {\n  this._v = v;\n}");
   });
 });
