+++ vitest.config.ts
@@ -1,0 +1,16 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    coverage: {
+      provider: "v8",
+      include: ["src/**/*.ts"],
+      exclude: [
+        "src/index.ts",
+        "src/types.ts",
+        "**/*.d.ts",
+        "test/**",
+      ],
+    },
+  },
+});
